+++ src/main/groovy/gql/ast/package-info.groovy
@@ -1,5 +1,0 @@
-/**
- * This package contains AST transformations processing public
- * annotations
- */
-package gql.ast